backend/pyproject.toml
[project]
name = "meteo-api"
version = "0.1.0"
description = "Weather Data API for InfoClimat - Data For Good Season 14"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.13",
    "djangorestframework>=3.15.0",
    "psycopg[binary]>=3.2.0",
    "django-environ>=0.11.0",
    "drf-spectacular>=0.27.0",
    "django-cors-headers>=4.4.0",
    "django-filter>=24.0",
    "gunicorn>=25.0.0",
    "whitenoise>=6.7.0",
    "numpy>=1.26.0",
    "django-extensions>=4.1",
    "django-prometheus>=2.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "pre-commit>=3.7.0",
    "ipython>=8.26.0",
    "django-extensions>=3.2.0",
    "factory-boy>=3.3.0",
     "pandas>=2.0",
]

notebooks = [
    "pandas>=2.0",
    "matplotlib>=3.8",
    "jupyterlab>=4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
    "N", # pep8-naming
]
ignore = [
    "E501",   # line too long — géré par le formatter
    "DJ001",  # null=True sur CharField — intentionnel pour les champs météo optionnels (HHMM)
]

[tool.ruff.lint.per-file-ignores]
# C420 : dict.fromkeys() — style non bloquant dans les tests
"weather/tests/**" = ["C4"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["_", "i", "j", "k"]

[tool.ruff.lint.isort]
known-first-party = ["config", "weather"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["config", "weather"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
